added condition to delete unresponsive file only if scan is complete

+++ domains/internal_pt.py
@@ -33,18 +33,26 @@ class InternalAssessment(DisplayHandler):
         """Lists all possible hosts on a network using ICMP protocol
          to increase your attack surface
         """
-        live_ip_count = self.network_manager.get_live_ips(output=self.output_file)
+        live_ip_count = self.network_manager.get_live_ips(
+            output=self.output_file)
         paths = self.network_manager.get_file_paths()
         live = f"{paths['live_hosts']}"
         unresponsive = f"{paths['unresponsive_hosts']}"
 
-        self.print_success_message(
-            message=f"Discovered {live_ip_count} hosts, and saved them to: ",
-            extras=live
-        )
-        # remove unresponsive hosts if scan completes uninterrupted
-        if unresponsive:
-         self.network_manager.remove_file(unresponsive)
+        if live:
+            self.print_success_message(
+                message=f"Discovered {live_ip_count} hosts, and saved them to: ",
+                extras=live
+            )
+        # Delete unresponsive file only if scan completed successfully
+        if self.network_manager.scan_complete and unresponsive:
+            self.network_manager.remove_file(unresponsive)
+            self.print_success_message(
+                "Scan completed, removed unresponsive file: ")
+        elif unresponsive:
+            self.print_warning_message(
+                "Scan incomplete, retaining unresponsive file ", file_path=unresponsive)
+       
 
     def netexec_module(self):
         # Using CrackmapExec / Netexec Module
